network: move wireless secret agent bits into a separate file
Related: rhbz#1847681
